Initial Commit
Ok this is a project that's spanned many years and doing an initial
commit at this stage is somewhat hilarious.  This started life as a
bunch of discreet parts screwed to a piece of plywood, and has
ultimately taken shape as a single cohesive board.

halloween-vending.yaml is my esphome configuration, should you want to
use something else it also has the appropriate pin definitions.

You'll need an ESP32 to drop onto the board to make this work.